Implementando o método post para adicionar objetos no banco de dados

+++ Cardapio/src/main/java/com/example/Cardapio/food/FoodResponseDTO.java
@@ -2,7 +2,7 @@ package com.example.Cardapio.food;
 
 import java.math.BigInteger;
 
-public record FoodResponseDTO(BigInteger id, String name, String title, Integer price){
+public record FoodResponseDTO(Integer id, String name, String title, Integer price){
     public FoodResponseDTO(Food food){
         this(food.getId(),
         food.getName(),
